+++ src/lib/components/LinksMenu.module.scss
@@ -1,10 +1,11 @@
-.mainNav {
+.linksMenuWrap {
   display: flex;
   align-items: center;
   gap: 32px;
 
   :global(.navLink) {
     position: relative;
+    z-index: 0;
 
     &:before {
       content: "";
@@ -16,5 +17,61 @@
       right: -16px;
       z-index: 1;
     }
+
+    & { // same as secondary styles
+      color: #AAAAAA;
+      font-family: 'Roboto';
+      font-weight: 500;
+      font-size: 14px;
+      line-height: 22px;
+      text-transform: uppercase;
+
+      &:hover, &:global(.hover) {
+        color: #FFFFFF;
+      }
+
+      &:global(.active) {
+        color: #82EACF;
+      }
+    }
+
+    &.primary {
+      font-family: 'Barlow Condensed';
+      font-size: 20px;
+      line-height: 24px;
+
+      &:hover, &:global(.hover), &:global(.active) {
+        color: #82EACF;
+      }
+    }
+
+    &.ternary {
+      font-size: 12px;
+      color: #767676;
+
+      &:hover, &:global(.hover) {
+        color: #0C0C0C;
+      }
+
+      &:global(.active) {
+        color: #0D664E;
+      }
+    }
+  }
+
+  :global(.navButton) {
+    font-family: 'Roboto';
+    font-style: normal;
+    font-weight: 700;
+    font-size: 12px;
+    line-height: 24px;
+    letter-spacing: 0.008em;
+    text-transform: uppercase;
+
+    color: #E9E9E9;
+    border: 1px solid #AAAAAA;
+    border-radius: 24px;
+
+    padding: 0 16px;
   }
 }
