+++ src/lib/components/TopNavbar.module.scss
@@ -1,0 +1,38 @@
+@import 'lib/styles/mixins.scss';
+
+.topNavbarWrap {
+  position: relative;
+  display: flex;
+  align-items: center;
+  background: #0C0C0C;
+  color: #AAAAAA;
+  padding: 18px 32px;
+
+  &.primary {
+    background: #0C0C0C;
+    color: #AAAAAA;
+  }
+
+  &.secondary {
+    background: #2A2A2A;
+    color: #AAAAAA;
+    padding: 12px 32px;
+  }
+
+  &.ternary {
+    background: #F4F4F4;
+    color: #AAAAAA;
+    padding: 12px 32px;
+  }
+
+  @include mobile {
+    height: 48px;
+    padding: 12px 16px;
+  }
+}
+
+.right {
+  flex: 1 0 auto;
+  display: flex;
+  justify-content: flex-end;
+}
