+++ src/lib/marketing-navigation/components/HoverMenu.module.scss
@@ -3,7 +3,6 @@
   top: 100%;
   left: 0;
   right: 0;
-  z-index: 9;
 
   background: #fff;
   box-shadow: 0px 6px 24px rgba(0, 0, 0, 0.1);
@@ -31,7 +30,7 @@
   :global(*:hover) ~ &, &:hover {
     opacity: 1;
     visibility: visible;
-    z-index: 1;
+    z-index: 19;
     transform: translateY(0);
     pointer-events: auto;
