--- src/lib/marketing-navigation/components/NavigationBar.module.scss
@@ -1,100 +1,0 @@
-.navbar {
-  &.navbar {
-    height: auto;
-  }
-}
-
-.navbar.primary {
-  background: #0C0C0C;
-  color: #AAAAAA;
-  padding: 18px 32px;
-
-  .mainNav {
-    :global(.navLink) {
-      color: #AAAAAA;
-      text-transform: uppercase;
-      font-family: 'Barlow Condensed';
-      font-style: normal;
-      font-weight: 500;
-      font-size: 20px;
-      line-height: 24px;
-
-      &:hover, &:global(.hover), &:global(.active) {
-        color: #82EACF;
-      }
-    }
-  }
-}
-
-.navbar.secondary {
-  background: #2A2A2A;
-  color: #AAAAAA;
-  padding: 12px 32px;
-
-  .mainNav {
-    :global(.navLink) {
-      color: #AAAAAA;
-      font-family: 'Roboto';
-      font-style: normal;
-      font-weight: 500;
-      font-size: 14px;
-      line-height: 22px;
-      text-transform: uppercase;
-
-      &:hover, &:global(.hover) {
-        color: #FFFFFF;
-      }
-
-      &:global(.active) {
-        color: #82EACF;
-      }
-    }
-
-    :global(.navButton) {
-      font-family: 'Roboto';
-      font-style: normal;
-      font-weight: 700;
-      font-size: 12px;
-      line-height: 24px;
-      letter-spacing: 0.008em;
-      text-transform: uppercase;
-
-      color: #E9E9E9;
-      border: 1px solid #AAAAAA;
-      border-radius: 24px;
-
-      padding: 0 16px;
-    }
-  }
-}
-
-.navbar.ternary {
-  background: #F4F4F4;
-  color: #AAAAAA;
-  padding: 12px 32px;
-
-  .mainNav {
-    :global(.navLink) {
-      font-family: 'Roboto';
-      font-style: normal;
-      font-weight: 500;
-      font-size: 12px;
-      line-height: 22px;
-      color: #767676;
-
-      &:hover, &:global(.hover) {
-        color: #0C0C0C;
-      }
-
-      &:global(.active) {
-        color: #0D664E;
-      }
-    }
-  }
-}
-
-.hoverMenu {
-  padding: 0;
-  top: 100%;
-  width: 100%;
-}
